# Reduce allocation on StreamingHubClient method calls