--- a/perf/Microbenchmark/Microbenchmark.Client/.gitignore
+++ b/perf/Microbenchmark/Microbenchmark.Client/.gitignore
@@ -1,0 +1,1 @@
+BenchmarkDotNet.Artifacts