--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f8d1cd17d285aa40aba05cb8861efa2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: