--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/LICENSE.TXT.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/LICENSE.TXT.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d91ceaa4a725879459c4908f7eb8a4f9
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: