--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d49c9170d8551e346b86aafb360e2fd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: