--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc8731f0410c3b1458d9b44a697880d2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: