--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.dll.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.dll.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 178d59b1c5291b64d9a206e568e79b9b
+labels:
+- NuGetForUnity
+PluginImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  iconMap: {}
+  executionOrder: {}
+  defineConstraints: []
+  isPreloaded: 0
+  isOverridable: 0
+  isExplicitlyReferenced: 0
+  validateReferences: 1
+  platformData:
+  - first:
+      Any: 
+    second:
+      enabled: 1
+      settings: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: