--- a/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.xml.meta
+++ b/src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3a999c3e0c15371408a73b75c174eeac
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: