… src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.dll
… src/MagicOnion.Client.Unity/Assets/Plugins/System.Threading.Channels/8.0.0/lib/netstandard2.1/System.Threading.Channels.xml
… tests/MagicOnion.Server.Tests/StreamingHubHandlerTest.cs